sales_interview_agent/requirements.txt
openai
chromadb
python-dotenv
sentence-transformers
tqdm